+++ src/DpressWebApp.php
@@ -42,6 +42,7 @@ use Dynart\Dpress\Plugin\PluginService;
 use Dynart\Dpress\Query\CoreQueries;
 use Dynart\Dpress\Query\QueryFactory;
 use Dynart\Dpress\Service\AuthService;
+use Dynart\Dpress\Theme\PageAssets;
 use Dynart\Dpress\Theme\ThemeService;
 
 /**
@@ -146,6 +147,7 @@ class DpressWebApp extends WebApp {
         DpressServices::registerWidgets(Micro::get(FormWidgets::class));
         DpressServices::registerShortcodes(Micro::get(Shortcodes::class));
         DpressServices::registerBlocks(Micro::get(Blocks::class));
+        DpressServices::registerPageAssets(Micro::get(PageAssets::class));
         $this->applyTheme($view);
 
         DpressServices::registerEntities(Micro::get(EntityManager::class));